+++ images/repmgr/entrypoint.sh
@@ -186,6 +186,7 @@ primary_safety_guard() {
 
 case "$SCRIPT_NAME" in
     "postgres"|"agent")
+        require_pg_bindir || exit 1
         export PATH=$PATH:$PG_BINDIR
         PGDATA=${PGDATA:-/var/lib/postgresql/data/pgdata}
         export PGDATA